Update fast phantomJs code to new selenium version
